Pendientes sapiens

+++ fuentes/sapiens/src/main/java/com/vortexbird/sapiens/service/ReporteService.java
@@ -4,10 +4,10 @@ public interface ReporteService {
 	
 	public String reportePruebaEstudiante(Integer prueId, Integer estuId) throws Exception;
 	
-	public String reportePruebaModulo(Long prueId) throws Exception;
+	public String reportePruebaModulo(Integer prueId) throws Exception;
 	
-	public String reportePruebaPrograma(Long prueId) throws Exception;
+	public String reportePruebaPrograma(Integer prueId) throws Exception;
 	
-	public String reportePruebaDetalleEstudiante(Long prueId) throws Exception;
+	public String reportePruebaDetalleEstudiante(Integer prueId) throws Exception;
 	
 }
